Fix scatter decompression recreation functions
The both the Decompress and Decompress2 scatter functions had an
off-by-one error. This was reported thanks to @domenukk and anonymous.
The Decompress2 also had an error with the backreference offset. This
was fixed by reexamining the __decompress2 function in some Shannon
firmware. To validate that everything was finally working, Ghidra
emulation of the scatterload functions was implemented and the resulting
output data hashed. Unfortunately Ghidra's emulation is too slow for
large scatter sizes, which means we are stuck with the error-prone
approach of signature searching for scatter operations.